feat(boogle): add game menu Closes #8
