+++ src/main.rs
@@ -2,7 +2,7 @@ pub mod enums;
 pub mod functions;
 pub mod game;
 pub mod loading_screen;
-pub mod menu;
+pub mod menus;
 pub mod structs;
 
 use crate::enums::GameState;
@@ -10,16 +10,16 @@ use crate::enums::GameState;
 use bevy::prelude::*;
 use bevy_framepace::{FramepacePlugin, FramepaceSettings, Limiter};
 
-use game::GamePlugin;
+use game::PongPlugin;
 use loading_screen::LoadingScreenPlugin;
-use menu::MenuPlugin;
+use menus::{game_menu::GameMenuPlugin, main_menu::MainMenuPlugin};
 
 fn main() {
     App::new()
         // setup Plugins
         .add_plugins(DefaultPlugins.set(WindowPlugin {
             primary_window: Some(Window {
-                title: "Ping Pong".to_string(),
+                title: "Games".to_string(),
                 // mode: WindowMode::Fullscreen(
                 //     MonitorSelection::Primary,
                 //     VideoModeSelection::Current,
@@ -34,8 +34,9 @@ fn main() {
             limiter: Limiter::from_framerate(60.0),
         })
         //rest for plugins
-        .add_plugins(MenuPlugin)
-        .add_plugins(GamePlugin)
+        .add_plugins(MainMenuPlugin)
+        .add_plugins(GameMenuPlugin)
+        .add_plugins(PongPlugin)
         .add_plugins(LoadingScreenPlugin)
         .init_state::<GameState>()
         .add_systems(Startup, setup)
