--- src/menu.rs
@@ -1,166 +1,0 @@
-use bevy::prelude::*;
-
-use crate::enums::GameState;
-use crate::functions::despawn_screen;
-
-const NORMAL_BUTTON: Color = Color::srgb(0.5, 0.5, 0.5);
-const HOVERED_BUTTON: Color = Color::srgb(0.9, 0.9, 0.9);
-
-/// all components on the MainScreen need this
-#[derive(Component)]
-struct OnMainScreen;
-
-/// button actions
-#[derive(Component)]
-enum MenuButtonAction {
-    Play,
-    Quit,
-}
-
-/// menustate
-#[derive(Clone, Copy, Default, Eq, PartialEq, Debug, Hash, States)]
-enum MenuState {
-    Main,
-    #[default]
-    Disabled,
-}
-
-/// menu plugin
-pub struct MenuPlugin;
-impl Plugin for MenuPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_state::<MenuState>()
-            .add_systems(OnEnter(GameState::Menu), menu_setup)
-            .add_systems(OnEnter(MenuState::Main), main_menu_setup)
-            .add_systems(OnExit(MenuState::Main), despawn_screen::<OnMainScreen>)
-            .add_systems(
-                Update,
-                (button_system, button_action).run_if(in_state(GameState::Menu)),
-            );
-    }
-}
-
-/// change GameState::Menu to MenuState::Main
-fn menu_setup(mut menu_state: ResMut<NextState<MenuState>>) {
-    menu_state.set(MenuState::Main);
-}
-
-/// spawns all elements for the mainmenu
-fn main_menu_setup(mut commands: Commands) {
-    let button_node = Node {
-        width: Val::Px(200.0),
-        height: Val::Px(65.0),
-        margin: UiRect::all(Val::Px(20.0)),
-        justify_content: JustifyContent::Center,
-        align_items: AlignItems::Center,
-        ..default()
-    };
-    let button_icon_node = Node {
-        width: Val::Px(30.0),
-        // This takes the icons out of the flexbox flow, to be positioned exactly
-        position_type: PositionType::Absolute,
-        // The icon will be close to the left border of the button
-        left: Val::Px(10.0),
-        ..default()
-    };
-    let button_text_font = TextFont {
-        font_size: 33.0,
-        ..default()
-    };
-
-    commands.spawn((
-        Node {
-            width: Val::Percent(100.0),
-            height: Val::Percent(100.0),
-            align_items: AlignItems::Center,
-            justify_content: JustifyContent::Center,
-            ..default()
-        },
-        OnMainScreen,
-        children![(
-            Node {
-                flex_direction: FlexDirection::Column,
-                align_items: AlignItems::Center,
-                ..default()
-            },
-            children![
-                (
-                    Text::new("Pong"),
-                    TextFont {
-                        font_size: 67.0,
-                        ..default()
-                    },
-                    //da muss sich noch was aendern
-                    Node {
-                        margin: UiRect::all(Val::Px(50.0)),
-                        ..default()
-                    },
-                ),
-                (
-                    Button,
-                    button_node.clone(),
-                    BackgroundColor(NORMAL_BUTTON),
-                    MenuButtonAction::Play,
-                    children![
-                        Text::new("New Game"),
-                        button_text_font.clone(),
-                        //man kann keine constanten hier verwenden
-                        button_icon_node.clone(),
-                    ]
-                ),
-                (
-                    Button,
-                    button_node.clone(),
-                    BackgroundColor(NORMAL_BUTTON),
-                    MenuButtonAction::Quit,
-                    children![
-                        button_icon_node.clone(),
-                        Text::new("Quit"),
-                        button_text_font.clone(),
-                    ]
-                )
-            ]
-        )],
-    ));
-}
-
-/// change the collors of the button if pressed or hovered
-fn button_system(
-    mut interaction_query: Query<
-        (&Interaction, &mut BackgroundColor),
-        (Changed<Interaction>, With<Button>),
-    >,
-) {
-    for (interaction, mut background_color) in &mut interaction_query {
-        *background_color = match *interaction {
-            Interaction::Pressed => Color::srgb(1.0, 1.0, 1.0).into(),
-            Interaction::Hovered => HOVERED_BUTTON.into(),
-            Interaction::None => NORMAL_BUTTON.into(),
-        }
-    }
-}
-
-/// change the state depends on wich button is pressed
-fn button_action(
-    interaction_query: Query<
-        (&Interaction, &MenuButtonAction),
-        (Changed<Interaction>, With<Button>),
-    >,
-    mut app_exit: EventWriter<AppExit>,
-    mut menu_state: ResMut<NextState<MenuState>>,
-    mut game_state: ResMut<NextState<GameState>>,
-) {
-    for (interaction, menu_button_action) in &interaction_query {
-        if *interaction == Interaction::Pressed {
-            match menu_button_action {
-                MenuButtonAction::Quit => {
-                    app_exit.write(AppExit::Success);
-                }
-                MenuButtonAction::Play => {
-                    game_state.set(GameState::Game);
-                    menu_state.set(MenuState::Disabled);
-                }
-            }
-        }
-    }
-}
